Preset and test setup for acceptance test transforms
