+++ lib/presets/acceptance.js
@@ -1,0 +1,3 @@
+const preset = require('./-preset');
+
+module.exports = preset('acceptance');
