+++ lib/transforms/integration/attr.js
@@ -1,4 +1,4 @@
-const { createFindExpression, createPropExpression, isJQuerySelectExpression, addImportStatement } = require('../utils');
+const { createFindExpression, createPropExpression, isJQuerySelectExpression, addImportStatement } = require('../../utils');
 
 /**
  * Creates a `find(selector).getAttribute(attr)` expression
