+++ lib/transforms/integration/click.js
@@ -1,7 +1,22 @@
-const { makeParentFunctionAsync, createTriggerExpression, isJQuerySelectExpression, addImportStatement } = require('../utils');
+const { migrateSelector, makeParentFunctionAsync, isJQuerySelectExpression, addImportStatement } = require('../../utils');
 
 /**
- * Check if `node` is a `this.$(selector).trigger(eventName)` expression
+ * Create `await click(selector)` expression
+ * @param j
+ * @param args
+ * @returns {*}
+ */
+function createExpression(j, args) {
+  return j.awaitExpression(
+    j.callExpression(
+      j.identifier('click'),
+      args.map(s => migrateSelector(j, s))
+    )
+  );
+}
+
+/**
+ * Check if `node` is a `this.$(selector).click()` expression
  *
  * @param j
  * @param node
@@ -12,13 +27,11 @@ function isJQueryExpression(j, node) {
     && j.MemberExpression.check(node.callee)
     && isJQuerySelectExpression(j, node.callee.object)
     && j.Identifier.check(node.callee.property)
-    && node.callee.property.name === 'trigger'
-    && node.arguments.length === 1
-    && j.Literal.check(node.arguments[0]);
+    && node.callee.property.name === 'click';
 }
 
 /**
- * Transforms `this.$(selector).trigger(eventName)` to `await triggerEvent(selector, eventName)`
+ * Transform `this.$(selector).click()` to `await click(selector)`
  *
  * @param file
  * @param api
@@ -33,12 +46,12 @@ function transform(file, api) {
   let replacements = root
     .find(j.CallExpression)
     .filter(({ node }) => isJQueryExpression(j, node))
-    .replaceWith(({ node }) => createTriggerExpression(j, node.callee.object.arguments[0], node.arguments[0]))
+    .replaceWith(({ node }) => createExpression(j, node.callee.object.arguments))
     .forEach((path) => makeParentFunctionAsync(j, path))
     ;
 
   if (replacements.length > 0) {
-    addImportStatement(j, root, ['triggerEvent']);
+    addImportStatement(j, root, ['click']);
   }
 
   return root.toSource({ quote: 'single' });
