+++ lib/transforms/integration/focus.js
@@ -1,22 +1,22 @@
-const { migrateSelector, makeParentFunctionAsync, isJQuerySelectExpression, addImportStatement } = require('../utils');
+const { migrateSelector, makeParentFunctionAsync, isJQuerySelectExpression, addImportStatement } = require('../../utils');
 
 /**
- * Create `await click(selector)` expression
+ * Create `await focus(selector)` expression
  * @param j
- * @param args
+ * @param selector
  * @returns {*}
  */
-function createExpression(j, args) {
+function createExpression(j, selector) {
   return j.awaitExpression(
     j.callExpression(
-      j.identifier('click'),
-      args.map(s => migrateSelector(j, s))
+      j.identifier('focus'),
+      migrateSelector(j, selector)
     )
   );
 }
 
 /**
- * Check if `node` is a `this.$(selector).click()` expression
+ * Check if `node` is a `this.$(selector).focus()` expression
  *
  * @param j
  * @param node
@@ -27,11 +27,11 @@ function isJQueryExpression(j, node) {
     && j.MemberExpression.check(node.callee)
     && isJQuerySelectExpression(j, node.callee.object)
     && j.Identifier.check(node.callee.property)
-    && node.callee.property.name === 'click';
+    && node.callee.property.name === 'focus';
 }
 
 /**
- * Transform `this.$(selector).click()` to `await click(selector)`
+ * Transform `this.$(selector).focus()` to `await focus(selector)`
  *
  * @param file
  * @param api
@@ -51,7 +51,7 @@ function transform(file, api) {
     ;
 
   if (replacements.length > 0) {
-    addImportStatement(j, root, ['click']);
+    addImportStatement(j, root, ['focus']);
   }
 
   return root.toSource({ quote: 'single' });
