+++ lib/transforms/integration/get-value.js
@@ -1,7 +1,7 @@
-const { createFindExpression, isJQuerySelectExpression, addImportStatement } = require('../utils');
+const { createFindExpression, isJQuerySelectExpression, addImportStatement } = require('../../utils');
 
 /**
- * Creates a `find(selector).textContent` expression
+ * Creates a `find(selector).value` expression
  *
  * @param j
  * @param findArgs
@@ -10,12 +10,12 @@ const { createFindExpression, isJQuerySelectExpression, addImportStatement } = r
 function createExpression(j, findArgs) {
   return j.memberExpression(
     createFindExpression(j, findArgs),
-    j.identifier('textContent')
+    j.identifier('value')
   );
 }
 
 /**
- * Check if `node` is a `this.$(selector).text()` expression
+ * Check if `node` is a `this.$(selector).val()` expression
  *
  * @param j
  * @param node
@@ -26,11 +26,12 @@ function isJQueryExpression(j, node) {
     && j.MemberExpression.check(node.callee)
     && isJQuerySelectExpression(j, node.callee.object)
     && j.Identifier.check(node.callee.property)
-    && node.callee.property.name === 'text';
+    && node.callee.property.name === 'val'
+    && node.arguments.length === 0;
 }
 
 /**
- * Transforms `this.$(selector).text()` to `find(selector).textContent`
+ * Transform `this.$(selector).val()` to `find(selector).value`
  *
  * @param file
  * @param api
