+++ lib/transforms/integration/has-class.js
@@ -1,4 +1,4 @@
-const { createFindExpression, isJQuerySelectExpression, addImportStatement } = require('../utils');
+const { createFindExpression, isJQuerySelectExpression, addImportStatement } = require('../../utils');
 
 /**
  * Creates a `find(selector).classList.contains(className)` expression
