+++ lib/transforms/integration/html.js
@@ -1,21 +1,22 @@
-const { createFindExpression, isJQuerySelectExpression, addImportStatement } = require('../utils');
+const { createFindExpression, isJQuerySelectExpression, addImportStatement } = require('../../utils');
 
 /**
- * Creates a `find(selector).value` expression
+ * Creates a `find(selector).innerHTML` expression
  *
  * @param j
  * @param findArgs
+ * @param attr
  * @returns {*}
  */
 function createExpression(j, findArgs) {
   return j.memberExpression(
     createFindExpression(j, findArgs),
-    j.identifier('value')
+    j.identifier('innerHTML')
   );
 }
 
 /**
- * Check if `node` is a `this.$(selector).val()` expression
+ * Check if `node` is a `this.$(selector).html()` expression
  *
  * @param j
  * @param node
@@ -26,12 +27,11 @@ function isJQueryExpression(j, node) {
     && j.MemberExpression.check(node.callee)
     && isJQuerySelectExpression(j, node.callee.object)
     && j.Identifier.check(node.callee.property)
-    && node.callee.property.name === 'val'
-    && node.arguments.length === 0;
+    && node.callee.property.name === 'html';
 }
 
 /**
- * Transform `this.$(selector).val()` to `find(selector).value`
+ * Transform `this.$(selector).html()` to `find(selector).innerHTML`
  *
  * @param file
  * @param api
