+++ lib/transforms/integration/key-event.js
@@ -1,4 +1,4 @@
-const { migrateSelector, makeParentFunctionAsync, isJQuerySelectExpression, addImportStatement } = require('../utils');
+const { migrateSelector, makeParentFunctionAsync, isJQuerySelectExpression, addImportStatement } = require('../../utils');
 
 /**
  * Creates a `await keyEvent(selector, eventName, keyCode)` expression
