+++ lib/transforms/integration/length.js
@@ -1,4 +1,4 @@
-const { createFindAllExpression, isJQuerySelectExpression, addImportStatement } = require('../utils');
+const { createFindAllExpression, isJQuerySelectExpression, addImportStatement } = require('../../utils');
 
 /**
  * Creates a `findAll(selector).length` expression
