+++ lib/transforms/integration/prop.js
@@ -1,4 +1,4 @@
-const { createPropExpression, isJQuerySelectExpression, addImportStatement } = require('../utils');
+const { createPropExpression, isJQuerySelectExpression, addImportStatement } = require('../../utils');
 
 /**
  * Check if `node` is a `this.$(selector).prop('propName')` expression
