+++ lib/transforms/integration/set-value.js
@@ -1,4 +1,4 @@
-const { migrateSelector, makeParentFunctionAsync, isJQuerySelectExpression, addImportStatement } = require('../utils');
+const { migrateSelector, makeParentFunctionAsync, isJQuerySelectExpression, addImportStatement } = require('../../utils');
 
 /**
  * Creates a `await fillIn(selector, value)` expression
