+++ lib/transforms/integration/text.js
@@ -1,22 +1,21 @@
-const { createFindExpression, isJQuerySelectExpression, addImportStatement } = require('../utils');
+const { createFindExpression, isJQuerySelectExpression, addImportStatement } = require('../../utils');
 
 /**
- * Creates a `find(selector).innerHTML` expression
+ * Creates a `find(selector).textContent` expression
  *
  * @param j
  * @param findArgs
- * @param attr
  * @returns {*}
  */
 function createExpression(j, findArgs) {
   return j.memberExpression(
     createFindExpression(j, findArgs),
-    j.identifier('innerHTML')
+    j.identifier('textContent')
   );
 }
 
 /**
- * Check if `node` is a `this.$(selector).html()` expression
+ * Check if `node` is a `this.$(selector).text()` expression
  *
  * @param j
  * @param node
@@ -27,11 +26,11 @@ function isJQueryExpression(j, node) {
     && j.MemberExpression.check(node.callee)
     && isJQuerySelectExpression(j, node.callee.object)
     && j.Identifier.check(node.callee.property)
-    && node.callee.property.name === 'html';
+    && node.callee.property.name === 'text';
 }
 
 /**
- * Transform `this.$(selector).html()` to `find(selector).innerHTML`
+ * Transforms `this.$(selector).text()` to `find(selector).textContent`
  *
  * @param file
  * @param api
