+++ lib/transforms/integration/trigger-shortcut.js
@@ -1,4 +1,4 @@
-const { makeParentFunctionAsync, createTriggerExpression, isJQuerySelectExpression, addImportStatement } = require('../utils');
+const { makeParentFunctionAsync, createTriggerExpression, isJQuerySelectExpression, addImportStatement } = require('../../utils');
 
 const triggerShortcuts = [
   'change',
