+++ lib/transforms/integration/trigger.js
@@ -1,22 +1,7 @@
-const { migrateSelector, makeParentFunctionAsync, isJQuerySelectExpression, addImportStatement } = require('../utils');
+const { makeParentFunctionAsync, createTriggerExpression, isJQuerySelectExpression, addImportStatement } = require('../../utils');
 
 /**
- * Create `await focus(selector)` expression
- * @param j
- * @param selector
- * @returns {*}
- */
-function createExpression(j, selector) {
-  return j.awaitExpression(
-    j.callExpression(
-      j.identifier('focus'),
-      migrateSelector(j, selector)
-    )
-  );
-}
-
-/**
- * Check if `node` is a `this.$(selector).focus()` expression
+ * Check if `node` is a `this.$(selector).trigger(eventName)` expression
  *
  * @param j
  * @param node
@@ -27,11 +12,13 @@ function isJQueryExpression(j, node) {
     && j.MemberExpression.check(node.callee)
     && isJQuerySelectExpression(j, node.callee.object)
     && j.Identifier.check(node.callee.property)
-    && node.callee.property.name === 'focus';
+    && node.callee.property.name === 'trigger'
+    && node.arguments.length === 1
+    && j.Literal.check(node.arguments[0]);
 }
 
 /**
- * Transform `this.$(selector).focus()` to `await focus(selector)`
+ * Transforms `this.$(selector).trigger(eventName)` to `await triggerEvent(selector, eventName)`
  *
  * @param file
  * @param api
@@ -46,12 +33,12 @@ function transform(file, api) {
   let replacements = root
     .find(j.CallExpression)
     .filter(({ node }) => isJQueryExpression(j, node))
-    .replaceWith(({ node }) => createExpression(j, node.callee.object.arguments))
+    .replaceWith(({ node }) => createTriggerExpression(j, node.callee.object.arguments[0], node.arguments[0]))
     .forEach((path) => makeParentFunctionAsync(j, path))
     ;
 
   if (replacements.length > 0) {
-    addImportStatement(j, root, ['focus']);
+    addImportStatement(j, root, ['triggerEvent']);
   }
 
   return root.toSource({ quote: 'single' });
