+++ test/integration-test.js
@@ -1,0 +1,3 @@
+const diff = require('./utils/diff-test');
+
+diff('integration');
