+++ test/integration/expected-output/all.js
