+++ test/integration/expected-output/attr.js
