+++ test/integration/expected-output/click.js
