+++ test/integration/expected-output/default-component-test.js
