+++ test/integration/expected-output/event.js
