+++ test/integration/expected-output/focus.js
