+++ test/integration/expected-output/has-class.js
