+++ test/integration/expected-output/jq-extensions.js
