+++ test/integration/expected-output/key-event.js
