+++ test/integration/expected-output/prop.js
