+++ test/integration/expected-output/text.js
