+++ test/integration/input/all.js
