+++ test/integration/input/attr.js
