+++ test/integration/input/click.js
