+++ test/integration/input/default-component-test.js
