+++ test/integration/input/focus.js
