+++ test/integration/input/get-value.js
