+++ test/integration/input/has-class.js
