+++ test/integration/input/jq-extensions.js
