+++ test/integration/input/key-event.js
