+++ test/integration/input/prop.js
