+++ test/integration/input/text.js
